Replace .reporting with genno.compat.ixmp as of v0.5.0; adjust imports
