+++ ixmp/reporting/util.py
@@ -1,0 +1,38 @@
+from functools import lru_cache
+from typing import Dict
+
+import pandas as pd
+
+#: Dimensions to rename when extracting raw data from Scenario objects.
+#: Mapping from Scenario dimension name -> preferred dimension name.
+RENAME_DIMS: Dict[str, str] = {}
+
+
+def dims_for_qty(data):
+    """Return the list of dimensions for *data*.
+
+    If *data* is a :class:`pandas.DataFrame`, its columns are processed;
+    otherwise it must be a list.
+
+    genno.RENAME_DIMS is used to rename dimensions.
+    """
+    if isinstance(data, pd.DataFrame):
+        # List of the dimensions
+        dims = data.columns.tolist()
+    else:
+        dims = list(data)
+
+    # Remove columns containing values or units; dimensions are the remainder
+    for col in "value", "lvl", "mrg", "unit":
+        try:
+            dims.remove(col)
+        except ValueError:
+            continue
+
+    # Rename dimensions
+    return [RENAME_DIMS.get(d, d) for d in dims]
+
+
+@lru_cache(1)
+def get_reversed_rename_dims():
+    return {v: k for k, v in RENAME_DIMS.items()}
